requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
marshmallow==3.14.1  # or any specific version you need
